% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Hypothesis Testing Function: myttest
The constructor will take at least three arguments x, y, and alpha and conducts hypothesis testing}
\usage{
myttest(x, y, alpha, paired = FALSE)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{alpha}{double}

\item{paired}{boolean}
}
\value{
a list object containing a dataframe of x and y, the alpha value, p-value and confidence interval
}
\description{
Hypothesis Testing Function: myttest
The constructor will take at least three arguments x, y, and alpha and conducts hypothesis testing
}
\examples{
set.seed(21); x <-rnorm(30,5,2); set.seed(23); y<- rnorm(40, 3,2); paired=FALSE
alpha <- 0.05; obj <- myttest(x = x, y = y, paired=FALSE, alpha = 0.05)

}
