% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Boxplots for myttest}
\usage{
plot(obj)
}
\arguments{
\item{obj}{Object of type Rttest, from myttest constructor}
}
\value{
Boxplot of differences if data is paired; Side by side boxplot if data is unpaired
}
\description{
Boxplots for myttest
}
\examples{
set.seed(21); x <-rnorm(30,5,2); set.seed(23); y<- rnorm(40, 3,2); paired=FALSE
alpha <- 0.05; obj <- myttest(x = x, y = y, paired=FALSE, alpha = 0.05); class(obj); plot(obj)

}
