% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Rttest Print Function}
\usage{
\method{print}{Rttest}(obj)
}
\arguments{
\item{obj}{Object of type Rttest, from myttest constructor}
}
\value{
The confidence interval and type of test performed
}
\description{
Rttest Print Function
}
\examples{
set.seed(21); x <-rnorm(30,5,2); set.seed(23); y<- rnorm(40, 3,2); paired=FALSE
alpha <- 0.05; obj <- myttest(x = x, y = y, paired=FALSE, alpha = 0.05); print(obj)

}
